+++ lib/color_observer_logger.dart
@@ -1,3 +1,4 @@
 export 'src/bloc_observer.dart';
 export 'src/ansi_color.dart';
 export 'src/logger_filter.dart';
+export 'src/bloc_hight_light_filter.dart';