+++ CHANGELOG.md
@@ -1,15 +1,16 @@
+## 0.0.9
+- Add BlocHightLightFilter design hight light message when something wrong 
+
+- Update logger print sequence
 ## 0.0.8
 
 - Update logger print sequence
-
 ## 0.0.7
 
 - Update README.md
-
 ## 0.0.6
 
 - Add LoggerFilter
-
 ## 0.0.4
 
 - Stable version